VectorDB/requirements.txt
chromadb==0.4.22
pymongo==4.6.1
sentence-transformers==2.2.2
numpy==1.26.3
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
tiktoken==0.5.2